Mini Boss has been created
This boss has 4 different states; the look is for looking player, the lock is for locking the player for a given time, and the dash is for main damage. while dashing if it hits the wall it will be stunned for a given time period afterward it will change its state into a look state.

Co-Authored-By: NurullahBaser <123994929+NurullahBaser@users.noreply.github.com>